requirements.txt
django==4.1.4
djangorestframework==3.11
psycopg2-binary==2.9.5
pytz==2022.7
argon2-cffi==21.2.0
django-filter==22.1
docker==6.0.1
black==22.12.0
django-simple-history==3.2.0